pyproject.toml
[project]
name = "rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "async-timeout==4.0.3",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.14.3",
    "bs4==0.0.2",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "emoji==2.15.0",
    "exceptiongroup==1.3.1",
    "faiss-cpu==1.13.2",
    "fastapi==0.128.0",
    "fastmcp>=2.14.0",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.45.0",
    "google-auth-httplib2==0.3.0",
    "google-genai==1.56.0",
    "google-generativeai==0.8.6",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.3.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "humanfriendly==10.0",
    "idna==3.11",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.5",
    "langchain-google-genai==4.1.2",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters==1.1.0",
    "langdetect==1.0.9",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.1",
    "lxml==6.0.2",
    "magika==0.6.3",
    "markdown-it-py==4.0.0",
    "markdownify==1.2.2",
    "markitdown==0.1.4",
    "marshmallow==3.26.2",
    "mcp[cli]>=1.25.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nltk==3.9.2",
    "numpy==2.2.6",
    "olefile==0.47",
    "onnxruntime==1.20.1",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "pandas==2.3.3",
    "propcache==0.4.1",
    "proto-plus==1.27.0",
    "protobuf==5.29.5",
    "psutil==7.2.1",
    "pyarrow==22.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pymilvus>=2.6.6",
    "pyparsing==3.3.1",
    "pypdf==6.5.0",
    "pyreadline3==3.5.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv==1.2.1",
    "python-iso639==2025.11.16",
    "python-magic==0.4.27",
    "python-oxmsg==0.0.2",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "rapidfuzz==3.14.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.1",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "unstructured==0.18.21",
    "unstructured-client==0.42.6",
    "uritemplate==4.2.0",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "uvicorn==0.40.0",
    "webencodings==0.5.1",
    "websockets==15.0.1",
    "wrapt==2.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
